Neue bzw. verbesserte Komponenten (#9)
* QS der Komponenten

* QS der Seiten

+++ app/client/src/main/svelte/pages/NutzerEditor.svelte
@@ -1,9 +1,12 @@
 <script>
     import { createEventDispatcher } from 'svelte';
+	import { toast } from '../components/Toast';
+	import { createValue } from '../utils/rest.js';
+	import { updatePatch } from '../utils/rest.js';
+	import { removeValue } from '../utils/rest.js';
 	import Button from '../components/Button';
-	import Dialog from '../components/Dialog';
-	import Groupbox from '../components/Groupbox';
 	import TextField from '../components/TextField';
+	import Toggle from '../components/Toggle';
     
     export let visible = false;
     export let nutzer = undefined;
@@ -18,14 +21,8 @@
         aktiv: true
     }
 
-	let allSpracheDialog = false;
-    function onAllSpracheDialogOk() {
-        allSpracheDialog = false;
-    }
-
-    $: disabled = !newNutzer.name || !newNutzer.mail || !newNutzer.allSprache;
-    $: if (nutzer) onChange()
-    function onChange() {
+    $: if (nutzer) onChangeNutzer()
+    function onChangeNutzer() {
         showUpdate = true;
         showRemove = !nutzer.aktiv;
         newNutzer = {
@@ -35,24 +32,50 @@
             allSprache: nutzer.allSprache,
             aktiv: nutzer.aktiv
         }
-        console.log(['onChange', newNutzer]);
+        console.log(['onChangeNutzer', newNutzer]);
     }
 
+    $: disabled = !newNutzer.name || !newNutzer.mail || !newNutzer.allSprache.length;
+
     const dispatch = createEventDispatcher();
-    function onCreate() {
-        visible = false;
-        console.log(['create', newNutzer]);
-        dispatch('create', newNutzer);
+    function onCreateNutzer() {
+        createValue('/api/nutzer', newNutzer)
+        .then(json => {
+            console.log(['onCreateNutzer', newNutzer, json]);
+            visible = false;
+            dispatch('create', newNutzer);
+        })
+        .catch(err => {
+            console.log(['onCreateNutzer', newNutzer, err]);
+            toast.push(err.toString());
+        });;
     }
-    function onUpdate() {
-        visible = false;
-        console.log(['update', newNutzer]);
-        dispatch('update', newNutzer);
+    function onUpdateNutzer() {
+        updatePatch('/api/nutzer' + '/' + newNutzer.id, newNutzer)
+        .then(json => {
+            console.log(['onUpdateNutzer', newNutzer, json]);
+            visible = false;
+            dispatch('update', newNutzer);
+        })
+        .catch(err => {
+            console.log(['onUpdateNutzer', newNutzer, err]);
+            toast.push(err.toString());
+        });;
     }
-    function onRemove() {
-        visible = false;
-        console.log(['remove', newNutzer]);
-        dispatch('remove', newNutzer);
+    function onRemoveNutzer() {
+        const text = newNutzer.name;
+        const hint = text.length > 20 ? text.substring(0, 20) + '...' : text;
+        if (!confirm("Nutzer '" + hint + "' wirklich löschen?")) return;
+        removeValue('/api/nutzer' + '/' + newNutzer.id)
+        .then(() => {
+            console.log(['onRemoveNutzer', newNutzer]);
+            visible = false;
+            dispatch('remove', newNutzer);
+        })
+        .catch(err => {
+            console.log(['onRemoveNutzer', newNutzer, err]);
+            toast.push(err.toString());
+        });;
     }
     function onCancel() {
         visible = false;
@@ -61,37 +84,38 @@
 
 <div class="flex flex-col">
     <div class="w-full">
-        <TextField bind:value={newNutzer.name} 
+        <TextField 
+            bind:value={newNutzer.name} 
             label="Name"		
             placeholder="Bitte den Namen eingeben"/>
     </div>
     <div class="w-full">
-        <TextField bind:value={newNutzer.mail}
+        <TextField 
+            bind:value={newNutzer.mail}
             label="E-Mail"		
             placeholder="Bitte die E-Mail-Adresse eingeben"/>
     </div>
-    <div class="grid grid-cols-1 lg:grid-cols-3 gap-2 py-2">
-        <Button on:click={() => allSpracheDialog=true}
-            disabeled={!newNutzer.name}
-            outlined block>
-            Sprachen
-            <small>({newNutzer.allSprache.length})</small>
-        </Button>
+    <div class="w-full">
+        <Toggle 
+            bind:allValue={newNutzer.allSprache}
+            allItem={allSpracheItem} 
+            label="Sprachen"
+            placeholder="Bitte Sprachen wählen"/>
     </div>
 </div>
 
 <div class="py-4">
     {#if showUpdate}
-    <Button on:click={() => onUpdate()} {disabled}>
+    <Button on:click={() => onUpdateNutzer()} {disabled}>
         Ok
     </Button>
     {:else}
-    <Button on:click={() => onCreate()} {disabled}>
+    <Button on:click={() => onCreateNutzer()} {disabled}>
         Ok
     </Button>
     {/if}
     {#if showRemove}
-    <Button on:click={() => onRemove()}>
+    <Button on:click={() => onRemoveNutzer()}>
         Löschen
     </Button>
     {/if}
@@ -100,21 +124,7 @@
     </Button>
 </div>
 
-<details>
+<details tabindex="-1">
     <summary>JSON</summary>
     <pre>{JSON.stringify(newNutzer, null, 2)}</pre>
 </details>
-
-<Dialog bind:value={allSpracheDialog}>
-	<h5 slot="title">
-		{newNutzer.name}s Sprachen
-	</h5>
-    <div class="flex flex-col gap-1">
-        <Groupbox 
-            bind:group={newNutzer.allSprache} 
-            items={allSpracheItem}/>
-    </div>
-    <div slot="actions">
-		<Button text on:click={onAllSpracheDialogOk}>Ok</Button>
-	</div>
-</Dialog>
